(fix) Move api handlers to server folder from db folder
