+++ server/api/user.js
@@ -1,0 +1,5 @@
+import model from "./db/models";
+
+const fetchUser = ()=>{
+
+};
