--- server/db/api/user.js
@@ -1,5 +1,0 @@
-import model from "../models";
-
-const fetchUser = ()=>{
-
-};
